track_mapper_flask/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9

